Installed Gems
This update installed multiple gems into the Gemfile. In addition i
separated production from development for some gems.

I included Oauth, an other dependent gems to allow sign in to all of
the major interface such as Facebook, google, and twitter. This should
cover most of the sign in needs. When couples with devise it will make
the sign in simple but robust.

I included mini magic to keep avatar files small, and simple form to
keep forms easy to code for. Carrierwave was installed for use in file
uploads to amazon s3, and faker for seed testing.